requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-appconf==1.0.6
django-cors-headers==4.3.1
django-filter==23.5
django-imagekit==5.0.0
django-json-widget==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
inflection==0.5.1
lxml==5.0.1
outcome==1.3.0.post0
packaging==23.2
pilkit==3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.17.2
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
transliterate==1.10.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wsproto==1.2.0
